sessionServer/requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
Django==5.1.2
django-apscheduler==0.6.2
django-celery-beat==2.7.0
django-jazzmin==3.0.1
django-oauth-toolkit==3.0.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
flower==2.0.1
hpack==4.0.0
httpcore==1.0.7
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
jwcrypto==1.5.6
kombu==5.4.2
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pyotp==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
redis==5.2.0
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
